Checked requirements and updated a ton of files
Now only parades with finalMode true and status accepted are visible to user not logged (we still need to do it for the members trying to create a request).

Added a method to list proclaims for a certain chapter to satisfy requirement 14.2.

The proclaim description can't be longer than 250 anymore.

Ordered by status the parades and fixed a problem within its list.